+++ internal/mithril-persistence/src/sqlite/mod.rs
@@ -5,6 +5,7 @@
 mod condition;
 mod connection_builder;
 mod connection_extensions;
+mod connection_pool;
 mod cursor;
 mod entity;
 mod projection;
@@ -15,6 +16,7 @@ mod transaction;
 pub use condition::{GetAllCondition, WhereCondition};
 pub use connection_builder::{ConnectionBuilder, ConnectionOptions};
 pub use connection_extensions::ConnectionExtensions;
+pub use connection_pool::{SqliteConnectionPool, SqlitePooledConnection};
 pub use cursor::EntityCursor;
 pub use entity::{HydrationError, SqLiteEntity};
 pub use projection::{Projection, ProjectionField};