+++ mithril-test-lab/mithril-end-to-end/src/mithril/aggregator.rs
@@ -86,6 +86,7 @@ impl Aggregator {
             ),
             ("CHAIN_OBSERVER_TYPE", aggregator_config.chain_observer_type),
             ("CARDANO_TRANSACTIONS_PROVER_CACHE_POOL_SIZE", "5"),
+            ("CARDANO_TRANSACTIONS_DATABASE_CONNECTION_POOL_SIZE", "5"),
             (
                 "CARDANO_TRANSACTIONS_SIGNING_CONFIG__SECURITY_PARAMETER",
                 "20",